🚨 chore(server): satisfy detekt on campfire surface
