+++ server/src/commonMain/kotlin/com/calypsan/listenup/server/plugins/AppErrorStatusPages.kt
@@ -101,7 +101,8 @@ fun Application.installAppErrorStatusPages() {
  * cyclomatic complexity under the project threshold of 25 while preserving compile-time
  * exhaustiveness: the client-local [InternalError]/[TransportError]/[PlaybackError] share a 500
  * branch, [ShelfError]/[SocialError] delegate to [shelfOrSocialHttpStatus], and
- * [LibraryError]/[LibraryWriteError] delegate to [libraryFamilyHttpStatus]. Adding a new
+ * [LibraryError]/[LibraryWriteError] delegate to [libraryFamilyHttpStatus], and
+ * [BookError]/[CoverError] delegate to [bookOrCoverHttpStatus]. Adding a new
  * [AppError] sub-interface will still fail this `when` at compile time.
  */
 internal fun AppError.toHttpStatus(): HttpStatusCode =
@@ -143,9 +144,10 @@ internal fun AppError.toHttpStatus(): HttpStatusCode =
 
         is InviteError -> toHttpStatus()
 
-        is BookError -> toHttpStatus()
-
-        is CoverError -> toHttpStatus()
+        // BookError + CoverError share one branch (delegating to an exhaustive helper) to keep
+        // this function's cyclomatic complexity under the project threshold while preserving
+        // per-variant exhaustiveness for both families.
+        is BookError, is CoverError -> bookOrCoverHttpStatus()
 
         is ContributorError -> toHttpStatus()
 
@@ -174,8 +176,9 @@ internal fun AppError.toHttpStatus(): HttpStatusCode =
  *
  * Exhaustive over all direct [AppError] implementors. The single-variant leaf families
  * ([ValidationError], [InternalError]) and the client-local types ([TransportError],
- * [PlaybackError]) are delegated to [leafWithCorrelationId] to keep the cyclomatic
- * complexity of this function under the project threshold of 25.
+ * [PlaybackError]) are delegated to [leafWithCorrelationId], and [ShelfError]/[SocialError]
+ * delegate to [shelfOrSocialWithCorrelationId], to keep the cyclomatic complexity of this
+ * function under the project threshold of 25.
  */
 internal fun AppError.withCorrelationId(id: String?): AppError =
     when (this) {
@@ -596,6 +599,18 @@ private fun MoodError.withCorrelationId(id: String?): MoodError =
         is MoodError.NameTooLong -> copy(correlationId = id)
     }
 
+/**
+ * Re-dispatches the grouped `BookError`/`CoverError` branch of [toHttpStatus] to each family's
+ * own exhaustive mapping. Split out solely to keep [toHttpStatus]'s cyclomatic complexity under
+ * the project threshold; the `else` is unreachable (only called from the grouped branch above).
+ */
+private fun AppError.bookOrCoverHttpStatus(): HttpStatusCode =
+    when (this) {
+        is BookError -> toHttpStatus()
+        is CoverError -> toHttpStatus()
+        else -> HttpStatusCode.InternalServerError // unreachable: only called from the grouped branch
+    }
+
 private fun BookError.toHttpStatus(): HttpStatusCode =
     when (this) {
         is BookError.NotFound -> HttpStatusCode.NotFound
@@ -722,10 +737,10 @@ private fun AppError.shelfOrSocialHttpStatus(): HttpStatusCode =
     }
 
 /**
- * Re-dispatches the grouped `ShelfError`/`SocialError`/`ReadingOrderError` branch of the
- * correlation-id stamp to each family's own exhaustive mapping. Split out solely to keep the
- * calling `when`'s cyclomatic complexity under the project threshold; the `else` is unreachable
- * (only called from the grouped branch above).
+ * Re-dispatches the grouped `ShelfError`/`SocialError`/`ReadingOrderError` branch of
+ * [withCorrelationId] to each family's own exhaustive mapping. Split out solely to keep
+ * [withCorrelationId]'s cyclomatic complexity under the project threshold; the `else` is
+ * unreachable (only called from the grouped branch above).
  */
 private fun AppError.shelfOrSocialWithCorrelationId(id: String?): AppError =
     when (this) {
