+++ server/src/jvmTest/kotlin/com/calypsan/listenup/server/api/CampfireServiceImplAccessTest.kt
@@ -2,41 +2,16 @@
 
 package com.calypsan.listenup.server.api
 
-import app.cash.sqldelight.db.SqlDriver
-import com.calypsan.listenup.api.dto.SharePermission
-import com.calypsan.listenup.api.dto.activity.ActivityType
-import com.calypsan.listenup.api.dto.auth.SessionId
-import com.calypsan.listenup.api.dto.auth.UserId
-import com.calypsan.listenup.api.dto.auth.UserRole
 import com.calypsan.listenup.api.dto.campfire.CampfireControlMode
 import com.calypsan.listenup.api.dto.campfire.CampfireFrame
 import com.calypsan.listenup.api.dto.campfire.CampfireSettings
 import com.calypsan.listenup.api.dto.campfire.PlaybackCommand
 import com.calypsan.listenup.api.error.CampfireError
-import com.calypsan.listenup.api.push.PushPayload
 import com.calypsan.listenup.api.result.AppResult
 import com.calypsan.listenup.api.streaming.RpcEvent
-import com.calypsan.listenup.api.sync.CollectionBookSyncPayload
-import com.calypsan.listenup.api.sync.CollectionShareSyncPayload
-import com.calypsan.listenup.api.sync.CollectionSyncPayload
 import com.calypsan.listenup.api.sync.SyncControl
-import com.calypsan.listenup.server.auth.PrincipalProvider
-import com.calypsan.listenup.server.auth.UserPrincipal
-import com.calypsan.listenup.server.auth.UserRoleLookup
-import com.calypsan.listenup.server.campfire.CampfireInviteNotifier
 import com.calypsan.listenup.server.campfire.CampfireRegistry
-import com.calypsan.listenup.server.db.sqldelight.ListenUpDatabase
-import com.calypsan.listenup.server.push.PushNotifier
-import com.calypsan.listenup.server.services.ActivityRecorder
-import com.calypsan.listenup.server.services.ActivityRepository
-import com.calypsan.listenup.server.services.ActivitySyncRepository
-import com.calypsan.listenup.server.services.PlaybackPositionRepository
 import com.calypsan.listenup.server.sync.ChangeBus
-import com.calypsan.listenup.server.sync.CollectionBookRepository
-import com.calypsan.listenup.server.sync.CollectionGrantRepository
-import com.calypsan.listenup.server.sync.CollectionRepository
-import com.calypsan.listenup.server.sync.PublicProfileRepository
-import com.calypsan.listenup.server.sync.SyncRegistry
 import com.calypsan.listenup.server.testing.FixedClock
 import com.calypsan.listenup.server.testing.seedTestBook
 import com.calypsan.listenup.server.testing.seedTestLibraryAndFolder
@@ -54,136 +29,22 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
 import kotlin.time.Duration.Companion.minutes
-import kotlin.time.Instant
 
 /**
  * Access + control gating tests for [CampfireServiceImpl] over a real, migrated in-memory
  * database (via [withSqlDatabase]) and a real [CampfireRegistry] — the [PushServiceImplTest]
  * pattern. Anchor math and pure room-state behavior are [CampfireRegistry]'s own tests
  * ([com.calypsan.listenup.server.campfire.CampfireRegistryTest]); these tests exercise the
- * layer above it: book-access gating, host-only enforcement, the §7 spoiler check, and the
- * observeSession presence contract.
+ * layer above it: book-access gating, host-only enforcement, the §7 spoiler check, the
+ * observeSession presence contract, and open-session discovery.
+ *
+ * Split from a single `CampfireServiceImplTest` (LargeClass) — invite/push-notification and
+ * "listened together" activity tests live in [CampfireServiceImplActivityTest]; shared fixtures
+ * (`makeService`, `principalFor`, `makeBookAccessible`, `seedPosition`/`seedChapter`, ...) live in
+ * `CampfireServiceImplTestFixtures.kt`.
  */
-class CampfireServiceImplTest :
+class CampfireServiceImplAccessTest :
     FunSpec({
-
-        val t0 = Instant.fromEpochMilliseconds(1_730_000_000_000L)
-        val everyoneSettings = CampfireSettings(controlMode = CampfireControlMode.EVERYONE, inviteOnly = false)
-        val hostOnlySettings = CampfireSettings(controlMode = CampfireControlMode.HOST_ONLY, inviteOnly = false)
-
-        fun principalFor(userId: String): PrincipalProvider =
-            PrincipalProvider { UserPrincipal(UserId(userId), SessionId("session-$userId"), UserRole.MEMBER) }
-
-        fun makeService(
-            sql: ListenUpDatabase,
-            driver: SqlDriver,
-            registry: CampfireRegistry,
-            principal: PrincipalProvider,
-            clock: kotlin.time.Clock = FixedClock(t0),
-            bus: ChangeBus = ChangeBus(),
-            pushNotifier: PushNotifier = RecordingPushNotifier(),
-            activityRecorder: ActivityRecorder =
-                ActivityRecorder(syncRepo = ActivitySyncRepository(db = sql, bus = bus, registry = SyncRegistry(), driver = driver)),
-        ): CampfireServiceImpl {
-            val syncRegistry = SyncRegistry()
-            return CampfireServiceImpl(
-                registry = registry,
-                bookAccessPolicy = BookAccessPolicy(sql, driver),
-                playbackPositions = PlaybackPositionRepository(db = sql, bus = bus, registry = syncRegistry),
-                publicProfiles = PublicProfileRepository(db = sql, bus = bus, registry = syncRegistry),
-                db = sql,
-                bus = bus,
-                userRoleLookup = UserRoleLookup(db = sql),
-                inviteNotifier = CampfireInviteNotifier(pushNotifier = pushNotifier),
-                activityRecorder = activityRecorder,
-                clock = clock,
-                principal = principal,
-            )
-        }
-
-        fun <T> AppResult<T>.value(): T {
-            this.shouldBeInstanceOf<AppResult.Success<T>>()
-            return data
-        }
-
-        /** Makes [bookId] visible to [viewer] via the ALL_BOOKS pure-union path (see [SocialServiceTest]). */
-        suspend fun makeBookAccessible(
-            sql: ListenUpDatabase,
-            driver: SqlDriver,
-            bookId: String,
-            viewer: String,
-            allBooksId: String = "all-books",
-        ) {
-            val bus = ChangeBus()
-            val syncRegistry = SyncRegistry()
-            val collectionRepo = CollectionRepository(db = sql, bus = bus, registry = syncRegistry, driver = driver)
-            val collectionBookRepo = CollectionBookRepository(db = sql, bus = bus, registry = syncRegistry, driver = driver)
-            val grantRepo = CollectionGrantRepository(db = sql, bus = bus, registry = syncRegistry, driver = driver)
-            collectionRepo.upsert(
-                CollectionSyncPayload(
-                    id = allBooksId,
-                    libraryId = "test-library",
-                    ownerId = "system",
-                    name = "All Books",
-                    isInbox = false,
-                    revision = 0L,
-                    updatedAt = 0L,
-                ),
-            )
-            collectionBookRepo.upsert(
-                CollectionBookSyncPayload(collectionId = allBooksId, bookId = bookId, createdAt = 0L, revision = 0L),
-            )
-            grantRepo.upsert(
-                CollectionShareSyncPayload(
-                    id = "grant-$viewer-$bookId",
-                    collectionId = allBooksId,
-                    sharedWithUserId = viewer,
-                    sharedByUserId = "system",
-                    permission = SharePermission.Read,
-                    revision = 0L,
-                    updatedAt = 0L,
-                ),
-            )
-        }
-
-        /** Directly inserts an unfinished playback position row for [userId]/[bookId]. */
-        fun ListenUpDatabase.seedPosition(
-            userId: String,
-            bookId: String,
-            positionMs: Long,
-        ) {
-            playbackPositionsQueries.insert(
-                id = "$userId-$bookId",
-                user_id = userId,
-                book_id = bookId,
-                position_ms = positionMs,
-                last_played_at = 1L,
-                finished = 0L,
-                playback_speed = 1.0,
-                current_chapter_id = null,
-                revision = 0L,
-                created_at = 1L,
-                updated_at = 1L,
-                deleted_at = null,
-                client_op_id = null,
-            )
-        }
-
-        /** Directly inserts a `book_chapters` row. */
-        fun ListenUpDatabase.seedChapter(
-            bookId: String,
-            ordinal: Long,
-            startTimeMs: Long,
-        ) {
-            bookChaptersQueries.insert(
-                book_id = bookId,
-                ordinal = ordinal,
-                id = "$bookId-ch-$ordinal",
-                title = "Chapter $ordinal",
-                duration = 0L,
-                start_time = startTimeMs,
-            )
-        }
 
         // ── Access gating ─────────────────────────────────────────────────────
 
@@ -193,9 +54,9 @@ class CampfireServiceImplTest :
                 sql.seedTestUser("member")
                 sql.seedTestBook("book-a")
                 runTest {
-                    val service = makeService(sql, driver, CampfireRegistry(clock = FixedClock(t0)), principalFor("member"))
+                    val service = makeService(sql, driver, CampfireRegistry(clock = FixedClock(campfireTestT0)), principalFor("member"))
 
-                    val result = service.createSession("book-a", everyoneSettings)
+                    val result = service.createSession("book-a", campfireEveryoneSettings)
 
                     result.shouldBeInstanceOf<AppResult.Failure>()
                     result.error.shouldBeInstanceOf<CampfireError.BookAccessDenied>()
@@ -211,8 +72,9 @@ class CampfireServiceImplTest :
                 sql.seedTestBook("book-a")
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
-                    val created = makeService(sql, driver, registry, principalFor("host")).createSession("book-a", everyoneSettings).value()
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
+                    val created =
+                        makeService(sql, driver, registry, principalFor("host")).createSession("book-a", campfireEveryoneSettings).value()
 
                     val result = makeService(sql, driver, registry, principalFor("stranger")).joinSession(created.id)
 
@@ -232,8 +94,8 @@ class CampfireServiceImplTest :
                     makeBookAccessible(sql, driver, "book-a", "host")
 
                     val snapshot =
-                        makeService(sql, driver, CampfireRegistry(clock = FixedClock(t0)), principalFor("host"))
-                            .createSession("book-a", everyoneSettings)
+                        makeService(sql, driver, CampfireRegistry(clock = FixedClock(campfireTestT0)), principalFor("host"))
+                            .createSession("book-a", campfireEveryoneSettings)
                             .value()
 
                     snapshot.anchor.positionMs shouldBe 123_000L
@@ -255,9 +117,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
                     // Room jumps 700s ahead of position 0 — past the 600s (10-minute) threshold.
                     hostService.sendCommand(created.id, PlaybackCommand.SeekTo(positionMs = 700_000L, commandId = "c1")).value()
 
@@ -282,9 +144,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
                     // 620s → chapter 3; within the 10-minute time threshold of 50s, but 3 chapters ahead.
                     hostService.sendCommand(created.id, PlaybackCommand.SeekTo(positionMs = 620_000L, commandId = "c1")).value()
 
@@ -305,9 +167,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
                     hostService.sendCommand(created.id, PlaybackCommand.SeekTo(positionMs = 200_000L, commandId = "c1")).value()
 
                     val joined = makeService(sql, driver, registry, principalFor("member")).joinSession(created.id).value()
@@ -328,9 +190,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
                     val memberService = makeService(sql, driver, registry, principalFor("member"))
                     memberService.joinSession(created.id).value()
 
@@ -351,9 +213,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
                     val memberService = makeService(sql, driver, registry, principalFor("member"))
                     memberService.joinSession(created.id).value()
 
@@ -374,9 +236,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
                     val memberService = makeService(sql, driver, registry, principalFor("member"))
                     memberService.joinSession(created.id).value()
 
@@ -397,9 +259,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", hostOnlySettings).value()
+                    val created = hostService.createSession("book-a", campfireHostOnlySettings).value()
                     val memberService = makeService(sql, driver, registry, principalFor("member"))
                     memberService.joinSession(created.id).value()
 
@@ -421,8 +283,9 @@ class CampfireServiceImplTest :
                 sql.seedTestBook("book-a")
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
-                    val created = makeService(sql, driver, registry, principalFor("host")).createSession("book-a", everyoneSettings).value()
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
+                    val created =
+                        makeService(sql, driver, registry, principalFor("host")).createSession("book-a", campfireEveryoneSettings).value()
 
                     val events = makeService(sql, driver, registry, principalFor("stranger")).observeSession(created.id).toList()
 
@@ -441,9 +304,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
                     val memberService = makeService(sql, driver, registry, principalFor("member"))
                     memberService.joinSession(created.id).value()
 
@@ -457,9 +320,13 @@ class CampfireServiceImplTest :
                     memberJob.cancel()
                     advanceUntilIdle()
 
-                    registry.snapshot(created.id)!!.members.first { it.userId == "member" }.isAway shouldBe true
+                    registry
+                        .snapshot(created.id)!!
+                        .members
+                        .first { it.userId == "member" }
+                        .isAway shouldBe true
 
-                    registry.reapAwayMembers(now = t0 + 3.minutes)
+                    registry.reapAwayMembers(now = campfireTestT0 + 3.minutes)
                     advanceUntilIdle()
 
                     observedFrames.filterIsInstance<CampfireFrame.MemberLeft>().any { it.member.userId == "member" } shouldBe true
@@ -482,10 +349,10 @@ class CampfireServiceImplTest :
                     makeBookAccessible(sql, driver, "book-a", "host")
                     makeBookAccessible(sql, driver, "book-a", "member")
                     makeBookAccessible(sql, driver, "book-b", "host", allBooksId = "all-books-b")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    hostService.createSession("book-a", everyoneSettings).value()
-                    hostService.createSession("book-b", everyoneSettings).value()
+                    hostService.createSession("book-a", campfireEveryoneSettings).value()
+                    hostService.createSession("book-b", campfireEveryoneSettings).value()
 
                     val visible = makeService(sql, driver, registry, principalFor("member")).listOpenSessions().value()
 
@@ -504,7 +371,7 @@ class CampfireServiceImplTest :
                 sql.seedTestBook("book-a")
                 runTest {
                     listOf("host", "member", "invited", "stranger").forEach { makeBookAccessible(sql, driver, "book-a", it) }
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val inviteOnlySettings =
                         CampfireSettings(
                             controlMode = CampfireControlMode.EVERYONE,
@@ -536,8 +403,8 @@ class CampfireServiceImplTest :
                     val sub = async { bus.subscribeControl().first().control }
                     advanceUntilIdle()
 
-                    makeService(sql, driver, CampfireRegistry(clock = FixedClock(t0)), principalFor("host"), bus = bus)
-                        .createSession("book-a", everyoneSettings)
+                    makeService(sql, driver, CampfireRegistry(clock = FixedClock(campfireTestT0)), principalFor("host"), bus = bus)
+                        .createSession("book-a", campfireEveryoneSettings)
                         .value()
 
                     sub.await() shouldBe SyncControl.CampfiresChanged
@@ -553,9 +420,9 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     val bus = ChangeBus()
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"), bus = bus)
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
 
                     val sub = async { bus.subscribeControl().first { it.control == SyncControl.CampfiresChanged }.control }
                     advanceUntilIdle()
@@ -574,191 +441,15 @@ class CampfireServiceImplTest :
                 runTest {
                     makeBookAccessible(sql, driver, "book-a", "host")
                     val bus = ChangeBus()
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
+                    val registry = CampfireRegistry(clock = FixedClock(campfireTestT0))
                     val hostService = makeService(sql, driver, registry, principalFor("host"), bus = bus)
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
+                    val created = hostService.createSession("book-a", campfireEveryoneSettings).value()
 
                     val sub = async { bus.subscribeControl().first { it.control == SyncControl.CampfiresChanged }.control }
                     advanceUntilIdle()
                     hostService.leaveSession(created.id).value()
 
                     sub.await() shouldBe SyncControl.CampfiresChanged
-                }
-            }
-        }
-
-        // ── Invites: push notification (Task 5) ────────────────────────────────
-
-        test(
-            "createSession push-notifies an invited user who can access the book, " +
-                "but never an inaccessible invitee or the inviter",
-        ) {
-            withSqlDatabase {
-                sql.seedTestLibraryAndFolder()
-                sql.seedTestUser("host")
-                sql.seedTestUser("member")
-                sql.seedTestUser("stranger")
-                sql.seedTestBook("book-a")
-                runTest {
-                    makeBookAccessible(sql, driver, "book-a", "host")
-                    makeBookAccessible(sql, driver, "book-a", "member")
-                    // "stranger" is deliberately never given access to "book-a".
-                    val notifier = RecordingPushNotifier()
-                    val inviteSettings =
-                        CampfireSettings(
-                            controlMode = CampfireControlMode.EVERYONE,
-                            inviteOnly = false,
-                            invitedUserIds = listOf("member", "stranger", "host"),
-                        )
-
-                    val created =
-                        makeService(
-                            sql,
-                            driver,
-                            CampfireRegistry(clock = FixedClock(t0)),
-                            principalFor("host"),
-                            pushNotifier = notifier,
-                        ).createSession("book-a", inviteSettings).value()
-
-                    notifier.calls.map { it.first } shouldBe listOf("member")
-                    val payload = notifier.calls.single().second.shouldBeInstanceOf<PushPayload.CampfireInvite>()
-                    payload.campfireId shouldBe created.id.value
-                    payload.bookId shouldBe "book-a"
-                    payload.inviterUserId shouldBe "host"
-                }
-            }
-        }
-
-        test("createSession with no invited users sends no push") {
-            withSqlDatabase {
-                sql.seedTestLibraryAndFolder()
-                sql.seedTestUser("host")
-                sql.seedTestBook("book-a")
-                runTest {
-                    makeBookAccessible(sql, driver, "book-a", "host")
-                    val notifier = RecordingPushNotifier()
-
-                    makeService(
-                        sql,
-                        driver,
-                        CampfireRegistry(clock = FixedClock(t0)),
-                        principalFor("host"),
-                        pushNotifier = notifier,
-                    ).createSession("book-a", everyoneSettings).value()
-
-                    notifier.calls.shouldBeEmpty()
-                }
-            }
-        }
-
-        // ── Invites: listInvitableUsers (Task 5) ───────────────────────────────
-
-        test("listInvitableUsers returns only active users who can access the book, excluding the caller") {
-            withSqlDatabase {
-                sql.seedTestLibraryAndFolder()
-                sql.seedTestUser("host")
-                sql.seedTestUser("member")
-                sql.seedTestUser("stranger")
-                sql.seedTestBook("book-a")
-                runTest {
-                    makeBookAccessible(sql, driver, "book-a", "host")
-                    makeBookAccessible(sql, driver, "book-a", "member")
-                    // "stranger" is deliberately never given access to "book-a".
-
-                    val invitable =
-                        makeService(sql, driver, CampfireRegistry(clock = FixedClock(t0)), principalFor("host"))
-                            .listInvitableUsers("book-a")
-                            .value()
-
-                    invitable.map { it.userId } shouldBe listOf("member")
-                }
-            }
-        }
-
-        // ── "Listened together" activity (Task 5) ──────────────────────────────
-
-        test("endSession by the host records CAMPFIRE_TOGETHER when at least 2 users ever joined") {
-            withSqlDatabase {
-                sql.seedTestLibraryAndFolder()
-                sql.seedTestUser("host")
-                sql.seedTestUser("member")
-                sql.seedTestBook("book-a")
-                runTest {
-                    makeBookAccessible(sql, driver, "book-a", "host")
-                    makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
-                    val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
-                    makeService(sql, driver, registry, principalFor("member")).joinSession(created.id).value()
-
-                    hostService.endSession(created.id).value()
-
-                    val rows = ActivityRepository(db = sql).page(before = null, limit = 10)
-                    rows shouldHaveSize 1
-                    rows.single().type shouldBe ActivityType.CAMPFIRE_TOGETHER
-                    rows.single().userId shouldBe "host"
-                    rows.single().bookId shouldBe "book-a"
-                }
-            }
-        }
-
-        test("endSession by a solo host does NOT record a listened-together activity") {
-            withSqlDatabase {
-                sql.seedTestLibraryAndFolder()
-                sql.seedTestUser("host")
-                sql.seedTestBook("book-a")
-                runTest {
-                    makeBookAccessible(sql, driver, "book-a", "host")
-                    val hostService =
-                        makeService(sql, driver, CampfireRegistry(clock = FixedClock(t0)), principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
-
-                    hostService.endSession(created.id).value()
-
-                    ActivityRepository(db = sql).page(before = null, limit = 10).shouldBeEmpty()
-                }
-            }
-        }
-
-        test("leaveSession by the last member ending a >=2-participant room records CAMPFIRE_TOGETHER") {
-            withSqlDatabase {
-                sql.seedTestLibraryAndFolder()
-                sql.seedTestUser("host")
-                sql.seedTestUser("member")
-                sql.seedTestBook("book-a")
-                runTest {
-                    makeBookAccessible(sql, driver, "book-a", "host")
-                    makeBookAccessible(sql, driver, "book-a", "member")
-                    val registry = CampfireRegistry(clock = FixedClock(t0))
-                    val hostService = makeService(sql, driver, registry, principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
-                    val memberService = makeService(sql, driver, registry, principalFor("member"))
-                    memberService.joinSession(created.id).value()
-                    memberService.leaveSession(created.id).value() // room continues — host still present
-
-                    hostService.leaveSession(created.id).value() // now empties and ends the room
-
-                    val rows = ActivityRepository(db = sql).page(before = null, limit = 10)
-                    rows shouldHaveSize 1
-                    rows.single().type shouldBe ActivityType.CAMPFIRE_TOGETHER
-                }
-            }
-        }
-
-        test("leaveSession by the last member of a solo room does NOT record a listened-together activity") {
-            withSqlDatabase {
-                sql.seedTestLibraryAndFolder()
-                sql.seedTestUser("host")
-                sql.seedTestBook("book-a")
-                runTest {
-                    makeBookAccessible(sql, driver, "book-a", "host")
-                    val hostService =
-                        makeService(sql, driver, CampfireRegistry(clock = FixedClock(t0)), principalFor("host"))
-                    val created = hostService.createSession("book-a", everyoneSettings).value()
-
-                    hostService.leaveSession(created.id).value()
-
-                    ActivityRepository(db = sql).page(before = null, limit = 10).shouldBeEmpty()
                 }
             }
         }
